3. After an album has finished go back to the main screen
Implemented this to return to screen from which user launched the album.
So if launched from main screen it returns there, but if launched from
an artist view showing albums for that artist, it returns to that
screen. This seems to be better than returning to the main screen in
that instance, but if you want it the other way I just need to add one
line; so let me know.